+++ src/parsers/mod.rs
@@ -92,6 +92,13 @@ pub trait Parser {
     fn parse_span(input: Span) -> ParseResult<Self::Item> {
         Self::parse(input)
     }
+
+    /// Uses the [`Parser::parse`] method to parse the input and, if successful,
+    /// discards the unparsed remaining input.
+    fn from_str(input: &str) -> Result<Self::Item, nom::Err<ParseError>> {
+        let (_, value) = Self::parse(input)?;
+        Ok(value)
+    }
 }
 
 /// Input type for parsers.